@@ -14,77 +14,94 @@
     "test": "mocha --recursive"
   },
   "dependencies": {
-    "animate.css": "^3.5.2",
+    "animate.css": "^3.6.1",
     "api.io": "^1.0.0",
     "async-busboy": "^0.6.2",
     "babel-core": "^6.18.2",
-    "babel-loader": "^7.1.1",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-async-to-module-method": "^6.16.0",
     "babel-polyfill": "^6.16.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-1": "^6.16.0",
     "bcryptjs": "^2.4.3",
     "bluebird": "^3.4.6",
-    "chart.js": "^2.2.2",
+    "chart.js": "^2.7.2",
     "checksum": "^0.1.1",
     "chron-time": "^0.0.2",
     "columnify": "^1.5.4",
-    "css-loader": "^0.28.8",
+    "compression-webpack-plugin": "^1.1.11",
+    "css-loader": "^0.28.11",
     "debounce": "^1.0.2",
-    "deepmerge": "^2.0.1",
+    "deepmerge": "^2.1.0",
     "emailjs": "^1.0.11",
     "expand-var": "^0.1.3",
     "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.6",
+    "file-loader": "^1.1.11",
     "file-uploader": "^1.0.0",
     "fs-extra": "^5.0.0",
-    "glob-promise": "^3.1.0",
-    "google-maps-react": "^1.1.0",
-    "immutability-helper": "^2.6.4",
-    "imports-loader": "^0.7.1",
+    "fs-extra-promise": "^1.0.1",
+    "glob-promise": "^3.4.0",
+    "google-maps-react": "^1.1.6",
+    "immutability-helper": "^2.6.6",
+    "imports-loader": "^0.8.0",
     "jsx-control-statements": "^3.1.5",
-    "koa": "^2.0.0",
+    "koa": "^2.5.0",
     "koa-bodyparser": "^4.1.0",
     "koa-compress": "^2.0.0",
     "koa-conditional-get": "^2.0.0",
     "koa-etag": "^3.0.0",
-    "koa-favicon": "^2.0.0",
+    "koa-favicon": "^2.0.1",
     "koa-range": "^0.3.0",
     "koa-route": "^3.2.0",
+    "koa-static": "^4.0.2",
     "koa-stream": "^0.2.0",
     "koa-webpack-dev-middleware": "^2.0.1",
-    "moment": "^2.18.1",
-    "moment-duration-format": "^2.1.1",
-    "mongodb": "^3.0.1",
+    "less": "^2.7.3",
+    "less-loader": "^4.0.5",
+    "less-plugin-rewrite-import": "^0.1.1",
+    "moment": "^2.21.0",
+    "moment-duration-format": "^2.2.2",
+    "mongodb": "^3.0.4",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-each": "^0.10.0",
+    "postcss-import": "^11.1.0",
+    "postcss-loader": "^2.1.3",
+    "postcss-mixins": "^6.2.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-router": "^4.1.2",
+    "react-notification": "^6.8.2",
+    "react-router-dom": "^4.2.0",
     "react-svg-loader": "^2.1.0",
-    "request": "^2.81.0",
+    "request": "^2.85.0",
     "request-promise-native": "^1.0.4",
+    "semantic-ui-css": "^2.3.1",
+    "semantic-ui-less": "^2.2.12",
+    "semantic-ui-react": "^0.79.0",
     "server-destroy": "^1.0.1",
     "sha1": "^1.1.1",
     "shuffle-array": "^1.0.1",
     "simple-xml2json": "^1.2.3",
-    "source-map-support": "^0.5.0",
+    "source-map-support": "^0.5.4",
     "strip-ansi": "^4.0.0",
-    "style-loader": "^0.19.1",
-    "uglify-js": "^3.3.7",
-    "url-loader": "^0.6.2",
-    "uuid": "^3.0.1",
+    "style-loader": "^0.20.3",
+    "uglify-js": "^3.3.16",
+    "url-loader": "^1.0.1",
+    "uuid": "^3.2.1",
     "walk-promise": "^0.2.0",
     "webpack": "^3.5.4",
     "wsh.js": "^0.0.8",
     "yargs": "^10.1.1"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.1",
-    "duplicate-package-checker-webpack-plugin": "^2.1.0",
-    "eslint": "^4.15.0",
+    "babel-eslint": "^8.2.2",
+    "duplicate-package-checker-webpack-plugin": "^3.0.0",
+    "eslint": "^4.19.1",
     "eslint-plugin-jsx-control-statements": "^2.2.0",
-    "eslint-plugin-react": "^7.2.0",
-    "mocha": "^4.1.0"
+    "eslint-plugin-react": "^7.7.0",
+    "mocha": "^5.0.4"
   },
   "engines": {
     "node": ">= 8.9.0"
