@@ -7,7 +7,7 @@ module.exports = async (session, abspath) => {
     const parentPath = path.dirname(abspath);
 
     const parent = await Node.resolve(session, parentPath);
-    const child = await Node.resolve(session, abspath);
+    const child = await Node.resolve(session, abspath, { nofollow: true });
 
     await parent.removeChild(session, child);
     await child.remove(session);
