@@ -3,7 +3,7 @@ import themeify from "lib/themeify";
 import theme from "./theme.css";
 
 const exports = {
-    Root: require("./Root")
+    Home: require("./Home")
 };
 
 module.exports = themeify(theme, exports);
